COMMERCE-6900 batch-engine-service - rename and apply changes to test classes
