+++ modules/apps/batch-engine/batch-engine-service/src/test/java/com/liferay/batch/engine/internal/writer/CSVBatchEngineExportTaskItemWriterImplTest.java
@@ -15,24 +15,29 @@
 package com.liferay.batch.engine.internal.writer;
 
 import com.liferay.petra.io.unsync.UnsyncByteArrayOutputStream;
+import com.liferay.petra.string.CharPool;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
+import com.liferay.petra.string.StringUtil;
 import com.liferay.portal.test.rule.LiferayUnitTestRule;
+
+import java.lang.reflect.Field;
 
 import java.util.Arrays;
 import java.util.Collections;
+import java.util.Date;
 import java.util.List;
+import java.util.Map;
 
+import org.junit.Assert;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
 
-import org.skyscreamer.jsonassert.JSONAssert;
-
 /**
  * @author Ivica Cardic
  */
-public class JSONBatchEngineExportTaskItemWriterTest
+public class CSVBatchEngineExportTaskItemWriterImplTest
 	extends BaseBatchEngineExportTaskItemWriterTestCase {
 
 	@ClassRule
@@ -48,44 +53,80 @@ public class JSONBatchEngineExportTaskItemWriterTest
 	@Test
 	public void testWriteRowsWithDefinedFieldNames2() throws Exception {
 		_testWriteRows(
-			Arrays.asList("createDate", "description", "id", "name"));
+			Arrays.asList(
+				"createDate", "description", "id", "name_en", "name_hr"));
 	}
 
 	@Test
 	public void testWriteRowsWithDefinedFieldNames3() throws Exception {
-		_testWriteRows(Arrays.asList("createDate", "id", "name"));
+		_testWriteRows(Arrays.asList("createDate", "id", "name_en"));
 	}
 
 	@Test
 	public void testWriteRowsWithDefinedFieldNames4() throws Exception {
 		_testWriteRows(
-			Arrays.asList("id", "name", "description", "createDate"));
+			Arrays.asList(
+				"id", "name_hr", "name_en", "description", "createDate"));
 	}
 
 	@Test
 	public void testWriteRowsWithEmptyFieldNames() throws Exception {
-		_testWriteRows(Collections.emptyList());
+		try {
+			_testWriteRows(Collections.emptyList());
+
+			Assert.fail();
+		}
+		catch (IllegalArgumentException illegalArgumentException) {
+		}
+	}
+
+	private String _formatValue(Object value) {
+		if (value == null) {
+			return StringPool.BLANK;
+		}
+
+		if (value instanceof Date) {
+			return dateFormat.format(value);
+		}
+
+		return value.toString();
 	}
 
 	private String _getExpectedContent(
-		List<String> fieldNames, List<Item> items) {
+			List<String> fieldNames, List<Item> items)
+		throws Exception {
 
 		StringBundler sb = new StringBundler();
 
-		if (fieldNames.isEmpty()) {
-			fieldNames = jsonFieldNames;
-		}
-
-		sb.append("[");
+		sb.append(StringUtil.merge(fieldNames, StringPool.COMMA));
+		sb.append(StringPool.NEW_LINE);
 
 		for (Item item : items) {
-			sb.append(getItemJSONContent(fieldNames, item));
-			sb.append(StringPool.COMMA);
+			for (String fieldName : fieldNames) {
+				int index = fieldName.indexOf(CharPool.UNDERLINE);
+
+				if (index == -1) {
+					Field field = fieldMap.get(fieldName);
+
+					sb.append(_formatValue(field.get(item)));
+				}
+				else {
+					Field field = fieldMap.get(fieldName.substring(0, index));
+
+					Map<?, ?> valueMap = (Map<?, ?>)field.get(item);
+
+					sb.append(
+						_formatValue(
+							valueMap.get(fieldName.substring(index + 1))));
+				}
+
+				sb.append(StringPool.COMMA);
+			}
+
+			sb.setIndex(sb.index() - 1);
+
+			sb.append(StringPool.NEW_LINE);
 		}
-
-		sb.setIndex(sb.index() - 1);
-
-		sb.append("]");
 
 		return sb.toString();
 	}
@@ -94,21 +135,22 @@ public class JSONBatchEngineExportTaskItemWriterTest
 		UnsyncByteArrayOutputStream unsyncByteArrayOutputStream =
 			new UnsyncByteArrayOutputStream();
 
-		try (JSONBatchEngineExportTaskItemWriter
-				jsonBatchEngineExportTaskItemWriter =
-					new JSONBatchEngineExportTaskItemWriter(
-						fieldMap.keySet(), fieldNames,
+		try (CSVBatchEngineExportTaskItemWriterImpl
+				csvBatchEngineExportTaskItemWriterImpl =
+					new CSVBatchEngineExportTaskItemWriterImpl(
+						StringPool.COMMA, fieldMap, fieldNames,
 						unsyncByteArrayOutputStream)) {
 
 			for (Item[] items : getItemGroups()) {
-				jsonBatchEngineExportTaskItemWriter.write(Arrays.asList(items));
+				csvBatchEngineExportTaskItemWriterImpl.write(
+					Arrays.asList(items));
 			}
 		}
 
 		String content = unsyncByteArrayOutputStream.toString();
 
-		JSONAssert.assertEquals(
-			_getExpectedContent(fieldNames, getItems()), content, true);
+		Assert.assertEquals(
+			_getExpectedContent(fieldNames, getItems()), content);
 	}
 
 }
