+++ modules/apps/batch-engine/batch-engine-service/src/test/java/com/liferay/batch/engine/internal/writer/JSONLBatchEngineExportTaskItemWriterImplTest.java
@@ -15,29 +15,24 @@
 package com.liferay.batch.engine.internal.writer;
 
 import com.liferay.petra.io.unsync.UnsyncByteArrayOutputStream;
-import com.liferay.petra.string.CharPool;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
-import com.liferay.petra.string.StringUtil;
 import com.liferay.portal.test.rule.LiferayUnitTestRule;
-
-import java.lang.reflect.Field;
 
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.Date;
 import java.util.List;
-import java.util.Map;
 
-import org.junit.Assert;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
 
+import org.skyscreamer.jsonassert.JSONAssert;
+
 /**
  * @author Ivica Cardic
  */
-public class CSVBatchEngineExportTaskItemWriterTest
+public class JSONLBatchEngineExportTaskItemWriterImplTest
 	extends BaseBatchEngineExportTaskItemWriterTestCase {
 
 	@ClassRule
@@ -53,78 +48,36 @@ public class CSVBatchEngineExportTaskItemWriterTest
 	@Test
 	public void testWriteRowsWithDefinedFieldNames2() throws Exception {
 		_testWriteRows(
-			Arrays.asList(
-				"createDate", "description", "id", "name_en", "name_hr"));
+			Arrays.asList("createDate", "description", "id", "name"));
 	}
 
 	@Test
 	public void testWriteRowsWithDefinedFieldNames3() throws Exception {
-		_testWriteRows(Arrays.asList("createDate", "id", "name_en"));
+		_testWriteRows(Arrays.asList("createDate", "id", "name"));
 	}
 
 	@Test
 	public void testWriteRowsWithDefinedFieldNames4() throws Exception {
 		_testWriteRows(
-			Arrays.asList(
-				"id", "name_hr", "name_en", "description", "createDate"));
+			Arrays.asList("id", "name", "description", "createDate"));
 	}
 
 	@Test
 	public void testWriteRowsWithEmptyFieldNames() throws Exception {
-		try {
-			_testWriteRows(Collections.emptyList());
-
-			Assert.fail();
-		}
-		catch (IllegalArgumentException illegalArgumentException) {
-		}
-	}
-
-	private String _formatValue(Object value) {
-		if (value == null) {
-			return StringPool.BLANK;
-		}
-
-		if (value instanceof Date) {
-			return dateFormat.format(value);
-		}
-
-		return value.toString();
+		_testWriteRows(Collections.emptyList());
 	}
 
 	private String _getExpectedContent(
-			List<String> fieldNames, List<Item> items)
-		throws Exception {
+		List<String> fieldNames, List<Item> items) {
 
 		StringBundler sb = new StringBundler();
 
-		sb.append(StringUtil.merge(fieldNames, StringPool.COMMA));
-		sb.append(StringPool.NEW_LINE);
+		if (fieldNames.isEmpty()) {
+			fieldNames = jsonFieldNames;
+		}
 
 		for (Item item : items) {
-			for (String fieldName : fieldNames) {
-				int index = fieldName.indexOf(CharPool.UNDERLINE);
-
-				if (index == -1) {
-					Field field = fieldMap.get(fieldName);
-
-					sb.append(_formatValue(field.get(item)));
-				}
-				else {
-					Field field = fieldMap.get(fieldName.substring(0, index));
-
-					Map<?, ?> valueMap = (Map<?, ?>)field.get(item);
-
-					sb.append(
-						_formatValue(
-							valueMap.get(fieldName.substring(index + 1))));
-				}
-
-				sb.append(StringPool.COMMA);
-			}
-
-			sb.setIndex(sb.index() - 1);
-
+			sb.append(getItemJSONContent(fieldNames, item));
 			sb.append(StringPool.NEW_LINE);
 		}
 
@@ -135,21 +88,22 @@ public class CSVBatchEngineExportTaskItemWriterTest
 		UnsyncByteArrayOutputStream unsyncByteArrayOutputStream =
 			new UnsyncByteArrayOutputStream();
 
-		try (CSVBatchEngineExportTaskItemWriter
-				csvBatchEngineExportTaskItemWriter =
-					new CSVBatchEngineExportTaskItemWriter(
-						StringPool.COMMA, fieldMap, fieldNames,
+		try (JSONLBatchEngineExportTaskItemWriterImpl
+				jsonlBatchEngineExportTaskItemWriterImpl =
+					new JSONLBatchEngineExportTaskItemWriterImpl(
+						fieldMap.keySet(), fieldNames,
 						unsyncByteArrayOutputStream)) {
 
 			for (Item[] items : getItemGroups()) {
-				csvBatchEngineExportTaskItemWriter.write(Arrays.asList(items));
+				jsonlBatchEngineExportTaskItemWriterImpl.write(
+					Arrays.asList(items));
 			}
 		}
 
 		String content = unsyncByteArrayOutputStream.toString();
 
-		Assert.assertEquals(
-			_getExpectedContent(fieldNames, getItems()), content);
+		JSONAssert.assertEquals(
+			_getExpectedContent(fieldNames, getItems()), content, true);
 	}
 
 }
